moved zobrist values to consts
